Renamed JSON files to match user ids, added data model for sleep data and implemented the controller with a GET API and the service to retrieve the sleep data

+++ src/main/java/org/eightsleep/model/User.java
@@ -4,8 +4,8 @@ import lombok.Value;
 
 @Value
 public class User {
-    int id;
-    int householdId;
+    Long id;
+    Long householdId;
     String name;
     Role role;
 }
